requirements/dev-requirements.txt
pytekukko==0.14.0

mypy==1.14.1
  homeassistant-stubs>=2024.4.0
  voluptuous-stubs>=0.1.1
ruff==0.9.3
